+++ packages/react-behave/src/components/Responsive/Responsive.js
@@ -1,29 +1,34 @@
 import debounce from 'lodash.debounce';
 import isEqual from 'lodash.isequal';
+import memoize from 'memoize-one';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import EventListener from 'react-event-listener';
-import renderMethod from '../../core/renderMethod';
+import ResizeObserver from '../ResizeObserver/ResizeObserver';
 
 // Corresponds to 10 frames at 60 Hz
 // Delay to next tick for tests.
 const RESIZE_INTERVAL = process.env.NODE_ENV === 'test' ? 0 : 166;
 
-function smallerOrEqualTo(screenSizes, refWidth, width) {
-  return screenSizes[width] <= screenSizes[refWidth];
+function smallerOrEqualTo(screenSizes, refScreenSize, screenSize) {
+  return screenSizes[screenSize] <= screenSizes[refScreenSize];
 }
 
-function biggerOrEqualTo(screenSizes, refWidth, width) {
-  return screenSizes[width] >= screenSizes[refWidth];
+function biggerOrEqualTo(screenSizes, refScreenSize, screenSize) {
+  return screenSizes[screenSize] >= screenSizes[refScreenSize];
 }
 
-function getScreenWidth(screenSizes, windowWidth) {
+function getScreenSize(screenSizes, width) {
   return Object.keys(screenSizes)
     .reverse()
-    .find(screenSizeName => windowWidth >= screenSizes[screenSizeName]);
+    .find(screenSizeName => width >= screenSizes[screenSizeName]);
 }
 
 /**
+ * [props-minimum]: #minimum-string-optional
+ * [props-maximum]: #maximum-string-optional
+ * [props-screensizes]: #screensizes-object-optional
+ *
  * Render content depending on the screen size.
  *
  * ## Usage
@@ -54,55 +59,87 @@ function getScreenWidth(screenSizes, windowWidth) {
 class Responsive extends Component {
   static propTypes = {
     /**
-     * Children to render.
-     * The children will only be rendered if the current width is satisfied by `props.maximum` and `props.minimum`.
-     * See the [Usage](#usage) for an example.
-     */
-    children: PropTypes.node,
-
-    /**
-     * Maximum screen width.
-     * Must be one of the keys of `props.screenSizes`.
-     */
-    maximum: PropTypes.string,
-
-    /**
-     * Minimum screen width.
-     * Must be one of the keys of `props.screenSizes`.
-     */
-    minimum: PropTypes.string,
-
-    /**
-     * _Parameters_: `width: String`
+     * _Parameters_: `screenSize: String`, `width: Number`
      *
-     * Render the content.
-     * This function will only be called if the current width is satisfied by `props.maximum` and `props.minimum`.
+     * Invoked to render the children.
      *
-     * Warning: The `props.children` takes precedence over `props.render` so don’t use both.
+     * The children will only be rendered if the current width is satisfied by [`props.maximum`][props-maximum] and [`props.minimum`][props-minimum].
      *
      * Example:
      *
      * ```jsx
+     * import React from 'react';
+     * import { Responsive } from 'react-behave';
+     *
      * const screenSizes = {
      *   // [...]
      * };
      *
      * function LargeScreen() {
      *   return (
-     *     <Responsive
-     *       minimum="lg"
-     *       screenSizes={screenSizes}
-     *       render={width => <p>The width is '{width}'</p>}
-     *     />
+     *     <Responsive minimum="lg" screenSizes={screenSizes}>
+     *       {screenSize => <p>The screen size is '{screenSize}'</p>}
+     *     </Responsive>
      *   );
      * }
      * ```
+     *
+     * For convenience, `props.children` can also be a React element.
+     * See [Usage](#usage).
      */
-    render: PropTypes.func,
+    children: PropTypes.oneOfType([PropTypes.func, PropTypes.node]).isRequired,
+
+    /**
+     * The target element used to compute the screen size.
+     *
+     * By default, `window.innerWidth` is used, but it is possible to change the target for local responsivity.
+     *
+     * Example:
+     *
+     * ```jsx
+     * import React, { Component } from 'react';
+     * import { Responsive } from 'react-behave';
+     *
+     * const screenSizes = {
+     *   // [...]
+     * };
+     *
+     * class LocallyResponsive extends Component {
+     *   ref = React.createRef();
+     *
+     *   render() {
+     *     return (
+     *       <section ref={this.ref}>
+     *         <Responsive target={this.ref} screenSizes={screenSizes}>
+     *           {(size, width) => (
+     *             <p>
+     *               The size is '{size}', the width is {width}px
+     *             </p>
+     *           )}
+     *         </Responsive>
+     *       </section>
+     *     );
+     *   }
+     * }
+     * ```
+     */
+    target: PropTypes.object,
+
+    /**
+     * Maximum screen width.
+     * Must be one of the keys of [`props.screenSizes`][props-screensizes].
+     */
+    maximum: PropTypes.string,
+
+    /**
+     * Minimum screen width.
+     * Must be one of the keys of [`props.screenSizes`][props-screensizes].
+     */
+    minimum: PropTypes.string,
 
     /**
      * The screen sizes to use.
-     * Each key is the name of a screen size that will be used by `props.maximum` and `props.minimum` props and the value is the starting width in pixels of this range.
+     * Each key is the name of a screen size that will be used by [`props.maximum`][props-maximum] and [`props.minimum`][props-minimum], and the value is the starting width in pixels of this range.
      *
      * For example:
      *
@@ -133,52 +170,58 @@ class Responsive extends Component {
     },
   };
 
-  state = {
-    screenSizes: null,
-    width: null,
-  };
-
-  static getDerivedStateFromProps(props, state) {
-    if (!isEqual(props.screenSizes, state.screenSizes)) {
-      return {
-        screenSizes: props.screenSizes,
-        width: getScreenWidth(props.screenSizes, window.innerWidth),
-      };
-    }
-    return null;
-  }
+  getScreenSize = memoize(getScreenSize, isEqual);
 
   componentWillUnmount() {
     this.handleResize.cancel();
   }
 
-  updateWidth = () => {
-    const width = getScreenWidth(this.props.screenSizes, window.innerWidth);
+  handleResize = debounce(() => this.forceUpdate(), RESIZE_INTERVAL);
 
-    if (width !== this.state.width) {
-      this.setState({ width });
-    }
-  };
-
-  handleResize = debounce(this.updateWidth, RESIZE_INTERVAL);
-
-  render() {
-    const { width } = this.state;
-    const { children, maximum, minimum, render, screenSizes } = this.props;
-
+  renderChildren(width) {
+    const screenSize = this.getScreenSize(this.props.screenSizes, width);
     let visible = true;
 
-    if (minimum) {
-      visible = biggerOrEqualTo(screenSizes, minimum, width);
+    if (this.props.minimum) {
+      visible = biggerOrEqualTo(
+        this.props.screenSizes,
+        this.props.minimum,
+        screenSize,
+      );
     }
 
-    if (visible && maximum) {
-      visible = smallerOrEqualTo(screenSizes, maximum, width);
+    if (visible && this.props.maximum) {
+      visible = smallerOrEqualTo(
+        this.props.screenSizes,
+        this.props.maximum,
+        screenSize,
+      );
+    }
+
+    if (visible) {
+      if (typeof this.props.children === 'function') {
+        return this.props.children(screenSize, width);
+      }
+
+      return this.props.children;
+    }
+
+    return null;
+  }
+
+  render() {
+    // Only use `window` if `props.target` is not specified in the props.
+    if (this.props.target !== undefined) {
+      return (
+        <ResizeObserver target={this.props.target}>
+          {({ width }) => this.renderChildren(width)}
+        </ResizeObserver>
+      );
     }
 
     return (
       <EventListener target="window" onResize={this.handleResize}>
-        {visible && renderMethod({ children, render }, width)}
+        {this.renderChildren(window.innerWidth)}
       </EventListener>
     );
   }