+++ packages/react-behave/src/components/Responsive/Responsive.stories.js
@@ -1,6 +1,7 @@
 import { object, select } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
+import { WithRef } from '../../../.storybook/components';
 import Responsive from './Responsive';
 
 const defaultScreenSizes = Responsive.defaultProps.screenSizes;
@@ -19,13 +20,14 @@ stories.add('Render function', () => {
       maximum={select('Maximum', screenSizesOptions, '')}
       minimum={select('Minimum', screenSizesOptions, '')}
       screenSizes={screenSizes}
-      render={width => (
+    >
+      {screenSize => (
         <div>
-          <h3>The width is '{width}'.</h3>
+          <h3>The screen size is '{screenSize}'.</h3>
           <em>Use the "Viewport" and "Knobs" tabs bellow to play around.</em>
         </div>
       )}
-    />
+    </Responsive>
   );
 });
 
@@ -33,4 +35,20 @@ stories.add('Render children', () => (
   <Responsive maximum="md" minimum="sm">
     <p>I'm only visible on 'sm' and 'md' screens</p>
   </Responsive>
+));
+
+stories.add('Local responsiveness', () => (
+  <WithRef>
+    {(ref, element) => (
+      <div ref={ref}>
+        <Responsive target={element}>
+          {(size, width) => (
+            <p>
+              The size is '{size}', width is {width}px
+            </p>
+          )}
+        </Responsive>
+      </div>
+    )}
+  </WithRef>
 ));